Code update
Axie Treasury v1.2 is the improved version of v1, v1 is the closest implementation of the proposed UI/dashboard presented in video format walkthrough. The document is meant to visualize the elements expected to be included in the proposed Axie Treasury dashboard. More functionality will be introduced in the coming updates.
Axie Treasury v2.4 is a minor update, Daily Bounty and Log In links are now floating to the right. No active link was included yet. Fonts were also updated. Awaiting finalization which version to move on with for further development.